% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedRIComputing.R
\name{adjustedRIComputing}
\alias{adjustedRIComputing}
\title{adjustedRIComputing: Compute the adjusted rand index}
\usage{
adjustedRIComputing(res, true.clusters)
}
\arguments{
\item{res}{Output of IntMultiOmics function}

\item{true.clusters}{A vector containing the true clusters}
}
\value{
the Adjusted Rand Index
}
\description{
adjustedRIComputing: Compute the adjusted rand index
}
\examples{
set.seed(333)
c_1 <- simulateY(J=1000, prop=0.1, noise=1)
c_2 <- simulateY(J=2000, prop=0.1, noise=1)
c_3 <- simulateY(J=500, prop=0.1,  noise=0.5)
data <- list(c_1$data, c_2$data, c_3$data)
res <- CrIMMix::IntMultiOmics(data, K=4, method="RGCCA") # ok
adjustedRIComputing(res,c_1$true.clusters)
}
