% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doICluster.R
\name{doiCluster}
\alias{doiCluster}
\title{doiCluster}
\usage{
doiCluster(
  data,
  K = 2,
  lambda = c(0.03, 0.03, 0.03, 0.03),
  type = rep("gaussian", length(data))
)
}
\arguments{
\item{data}{List of matrices.}

\item{K}{Number of clusters}

\item{lambda}{Vector of lasso penalty terms.}

\item{type}{Data type, which can be gaussian, binomial, poisson, multinomial.}
}
\value{
a list of \code{clust} the clustering of samples and
\code{fit} the results of the method iCluster
}
\description{
doiCluster
}
\examples{
set.seed(333)
c_1 <- simulateY(J=100, prop=0.1, noise=1)
c_2 <- simulateY(J=200, prop=0.1, noise=1)
c_3 <- simulateY(J=50, prop=0.1,  noise=0.5)
data <- list(c_1$data , c_2$data , c_3$data)
\donttest{res <- doiCluster(data,K=4)}
}
