% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doIntNMF.R
\name{doIntNMF}
\alias{doIntNMF}
\title{doIntNMF}
\usage{
doIntNMF(data, K)
}
\arguments{
\item{data}{List of matrices.}

\item{K}{Number of latent variables and number of clusters}
}
\value{
a list of \code{clust} the clustering of samples and
\code{fit} the results of the method intNMF
}
\description{
doIntNMF
}
\examples{
set.seed(33)
c_1 <- simulateY(J=100, prop=0.1, noise=1)
c_2 <- simulateY(J=200, prop=0.1, noise=1)
c_3 <- simulateY(J=50, prop=0.1,  noise=0.5)
data <- list(c_1$data , c_2$data , c_3$data)
\donttest{res <- doIntNMF(data,K=4)}
}
