% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doLRAcluster.R
\name{doLRAcluster}
\alias{doLRAcluster}
\title{doLRAcluster}
\usage{
doLRAcluster(data, K, type = rep("gaussian", length(data)))
}
\arguments{
\item{data}{List of matrices.}

\item{K}{Number of clusters}

\item{type}{type of data ("binary","gaussian","poisson") by default gaussian}
}
\value{
a list of \code{clust} the clustering of samples and
\code{fit} the results of the method LRACluster
}
\description{
doLRAcluster
}
\examples{
set.seed(333)
c_1 <- simulateY(J=1000, prop=0.1, noise=1)
c_2 <- simulateY(J=2000, prop=0.1, noise=1)
c_3 <- simulateY(J=500, prop=0.1,  noise=0.5)
data <- list(c_1$data , c_2$data , c_3$data)
res <- doSNF(data,K=4, K_n=10, sigma=0.5)
}
