% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMCIA.R
\name{doMCIA}
\alias{doMCIA}
\title{doMCIA}
\usage{
doMCIA(data, K, ncomp = 10)
}
\arguments{
\item{data}{List of matrices.}

\item{K}{Number of clusters}

\item{ncomp}{Number of components}
}
\value{
a list of \code{clust} the clustering of samples and
\code{fit} the results of the method MCIA
}
\description{
doMCIA
}
\examples{
set.seed(333)
c_1 <- simulateY(J=100, prop=0.1, noise=1)
c_2 <- simulateY(J=200, prop=0.1, noise=1)
c_3 <- simulateY(J=50, prop=0.1,  noise=0.5)
data <- list(c_1$data , c_2$data , c_3$data)
res <- doMCIA(data,K=4,ncomp=3)
}
