% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMixKernel.R
\name{doMixKernel}
\alias{doMixKernel}
\title{doMixKernel}
\usage{
doMixKernel(data, K, kernel.func = "linear", ncomp = 2)
}
\arguments{
\item{data}{List of matrices.}

\item{K}{Number of clusters}

\item{kernel.func}{the kernel function to use.
This parameter can be set to any user defined kernel function.
Widely used kernel functions are pre-implemented, that can be used by setting kernel.func
to one of the following strings: "kidentity", "abundance",
"linear" or "phylogenetic". Default: "linear".}

\item{ncomp}{integer. Indicates the number of components to return.}
}
\value{
a list of \code{clust} the clustering of samples and
\code{fit} the results of the method intNMF#' @export
}
\description{
doMixKernel
}
\examples{
set.seed(333)
c_1 <- simulateY(J=100, prop=0.1, noise=1)
c_2 <- simulateY(J=200, prop=0.1, noise=1)
c_3 <- simulateY(J=50, prop=0.1,  noise=0.5)
data <- list(c_1$data , c_2$data , c_3$data)
res <- doMixKernel(data,K=4)
}
