% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMOFA.R
\name{doMOFA}
\alias{doMOFA}
\title{doMOFA}
\usage{
doMOFA(data, K)
}
\arguments{
\item{data}{List of matrices.}

\item{K}{Number of clusters}
}
\value{
a list of \code{clust} the clustering of samples and
\code{fit} the results of the method MOFA
}
\description{
doMOFA
}
\examples{
\dontrun{c_1 <- simulateY(J=100, prop=0.1, noise=1)
c_2 <- simulateY(J=200, prop=0.1, noise=1)
c_3 <- simulateY(J=50, prop=0.1,  noise=0.5)
data <- list(c_1$data , c_2$data , c_3$data)
res <- doMOFA(data,K=4)}
}
