% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doPINSPLUS.R
\name{doPINSPlus}
\alias{doPINSPlus}
\title{doPINSPlus}
\usage{
doPINSPlus(data, K)
}
\arguments{
\item{data}{List of matrices.}

\item{K}{Number of clusters}
}
\value{
a list of \code{clust} the clustering of samples and
\code{fit} the results of the method PINSPLUS
}
\description{
doPINSPlus
}
\examples{
set.seed(333)
c_1 <- simulateY(J=1000, prop=0.1, noise=1)
c_2 <- simulateY(J=2000, prop=0.1, noise=1)
c_3 <- simulateY(J=500, prop=0.1,  noise=0.5)
data <- list(c_1$data , c_2$data , c_3$data)
res <- doPINSPlus(data,K=4, K_n=10, sigma=0.5)
}
