% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRGCCA.R
\name{doRGCCA}
\alias{doRGCCA}
\title{doRGCCA}
\usage{
doRGCCA(data, K, C = 1 - diag(length(data)), ncomp = rep(1,
  length(data)), scheme = "centroid")
}
\arguments{
\item{data}{List of matrices.}

\item{K}{Nmber of clusters}

\item{C}{A design matrix that describes the relationships between blocks (default: complete design).}

\item{ncomp}{A 1 * J vector that contains the numbers of c
omponents for each block (default: \code{rep(1, length(data}),
which gives one component per block.)}

\item{scheme}{The value is "horst", "factorial", "centroid"
or any diffentiable convex scheme function g designed by
the user (default: "centroid").}
}
\value{
a list of \code{clust} the clustering of samples and
\code{fit} the results of the method RGCCA
}
\description{
doRGCCA
}
\examples{
set.seed(333)
c_1 <- simulateY(J=100, prop=0.1, noise=1)
c_2 <- simulateY(J=200, prop=0.1, noise=1)
c_3 <- simulateY(J=50, prop=0.1,  noise=0.5)
data <- list(c_1$data , c_2$data , c_3$data)
res <- doRGCCA(data,K=4)
}
