% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPos.R
\name{extractPos}
\alias{extractPos}
\title{Extract positive biomarkers}
\usage{
extractPos(fit, method)
}
\arguments{
\item{fit}{the result of the method}

\item{method}{method used}
}
\value{
selected Variables
}
\description{
Extract positive biomarkers
}
