% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_i.R
\name{remove_i}
\alias{remove_i}
\title{Remove observation i from data}
\usage{
remove_i(data, i)
}
\arguments{
\item{data}{List of matrices.}

\item{i}{observation to remove}
}
\value{
data set without observation i
}
\description{
Remove observation i from data
}
