% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_eval.R
\name{roc_eval}
\alias{roc_eval}
\title{roc_eval computation}
\usage{
roc_eval(truth, fit, method)
}
\arguments{
\item{truth}{a list of biomarkers for each data set}

\item{fit}{the result of the method}

\item{method}{method used}
}
\value{
roc_eval (TPR and FPR) for each data set
}
\description{
roc_eval computation
}
