% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateY}
\alias{simulateY}
\title{Simulate various type of data}
\usage{
simulateY(nclust = 4, n_byClust = 20, J, prop = 0.01, noise = 0.1,
  flavor = c("normal", "beta", "binary"), params = list(c(mean = 1, sd
  = 1)))
}
\arguments{
\item{nclust}{A numeric value specifying the number of cluster (integer)}

\item{n_byClust}{A list or numeric value specifying the number of patients by cluster (if only one value is specify groups contain the same number of patients)}

\item{J}{A numeric value specifying the number of total biomarkers}

\item{prop}{A numeric value specifying the proportion of informative biomarkers for each cluster}

\item{noise}{A numeric value specifying the noise for back ground distribution}

\item{flavor}{A character value specifying the distribution of simulations
Defaults to \code{"normal"}. See Details}

\item{params}{A list of the parameters for flavor (See Details)}
}
\value{
A list with two components: \describe{ \item{data}{A \code{nclust*n_byclust} x
  \code{J} matrix}, \item{positive}{the features that drives clusters} }
}
\description{
simulateY
}
\details{
The observations are simulated as follows according to the
  value of argument 'flavor':

  If \code{flavor=="normal"} (the default), the observations follow a normal distribution (specify in \code{params} the mean and the standard deviance for each cluster.
  \code{params} is a list of vector \code{list(c(mean, sd))}. If the length is one we
  consider that parameters are the same for all groups, else the length of the list needs
  to be of length \code{nclust}

  If \code{flavor=="beta"}, the observations follow a beta distribution (methylation data).

  If \code{flavor=="binary"}, the observations follow a bernoulli (0 or 1) distribution (specify in \code{params} the proportion of 1). For example to simulate mutations
  \code{params} is a list of vector \code{list(c(p))}. If the length is one we
  consider that parameters are the same for all groups, else the length of the list needs
  to be of length \code{nclust}
}
\examples{
c <- simulateY(J=100, prop=0.1)
heatmap(c$data, scale="none")
c_bin <- simulateY(J=100, flavor="binary",
params=list(c(p=0.7)), prop=0.1, noise=0.1)
heatmap(c_bin$data, scale="none")
c_beta <- simulateY(J=400, flavor="beta",
params=list(c(mean1=-3,mean2=3, sd1=0.1, sd2=0.2) ), prop=0.3, noise=1)
heatmap(c_beta$data[,unlist(c_beta$positive)], scale="none")
}
