% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doConsensusClustering.R
\name{doConsensusClustering}
\alias{doConsensusClustering}
\title{doConsensusClustering}
\usage{
doConsensusClustering(data, K, K_n = 10, sigma = 0.5)
}
\arguments{
\item{data}{List of matrices.}

\item{K}{Number of clusters}

\item{K_n}{Number of nearest neighbors}

\item{sigma}{Variance for local model}
}
\value{
a list of \code{clust} the clustering of samples and
\code{fit} the results of the method Consensus Clustering
}
\description{
doConsensusClustering
}
\examples{
set.seed(333)
c_1 <- simulateY(J=1000, prop=0.1, noise=1)
c_2 <- simulateY(J=2000, prop=0.1, noise=1)
c_3 <- simulateY(J=500, prop=0.1,  noise=0.5)
data <- list(c_1$data , c_2$data , c_3$data)
res <- doSNF(data,K=4, K_n=10, sigma=0.5)
}
